rename legacy_shortcodes_data_fix to link_unescape
